角色等级与数值机制大幅更新

+++ Assets/StagesManagement/FightInfo.cs
@@ -40,7 +40,7 @@ public class FightInfo : ScriptableObject
     public AIMode team1AIMode = AIMode.Aggressive;
     public AIMode team2AIMode = AIMode.Aggressive;
     public int dumbAIDecisionDelay = 20;
-    public int stageRefLevel = 1;
+    public float stageRefLevel = 1;
     
     public int ArcadeFightMode
     {
